#2 Refactors CricketLaneView into its own file
